Add compatibility shim for the original `mkvimball`
For now, `mkvimball-sh` hews pretty closely to the original `mkvimball`
(<https://www.drchip.org/astronaut/src/index.html#MKVIMBALL>) in terms
of how it processes its arguments, but it will diverge a bit as I add
functionality that brings it more in line with standard Unix tools.  My
plans include a few options, recognition of "--" as an end-of-options
marker, and interpretation of "-" operands to indicate standard input
and output.  Real cutting-edge stuff.

As a hedge, implement a compatibility shim that will continue to closely
imitate `mkvimball`.  Such imitation currently includes:

  - Rejecting output paths that contain "." characters.
  - Appending the deprecated ".vba" suffix to accepted output paths.
  - Interpreting future options, "--", and "-" as regular pathnames.