Teach mkvimball-sh to write archive to stdout
In emulation of pax(1) [1], the archive path is now passed using -f,
which is optional; if -f <archive> is not specified, mkvimball-sh writes
the vimball archive to standard output.  Passing the archive path with
an option allows omitting it independently of the input paths [2].

Now that mkvimball-sh recognizes an option, it also recognizes "--" to
indicate the end of options.

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/pax.html
[2]: 267e287 (Teach mkvimball-sh to read input paths from stdin,
     2022-03-08)