man: Remove `docyear` from "Copyright" section
The `docyear` attribute only provides the year in which the AsciiDoc
document was last modified; this obviously omits all the other years in
which content was written.  But I don't really want to replace it with
another list of years that I have to keep updated.

Additionally, authorship dates are ambiguous in this context.  Does
"Written in 2022-2023" refer to the manual itself or to the software?
There's no difference right now, but there could be in the future, and
I honestly can't be bothered to include clarification.

Just remove the date altogether.  There are many other places in this
repository where dates of interest can be found.